split page and pages
