+++ src/lib/page.ts
@@ -1,0 +1,89 @@
+import { fabric } from "fabric";
+
+export class Page extends fabric.Canvas {
+  cursor: { x: number; y: number };
+  latestId: number = 0;
+  modified: boolean = false;
+
+  fitToWindow = async (
+    canvasWidth: number,
+    canvasHeight: number
+  ): Promise<void> => {
+    const widthRatio = window.innerWidth / canvasWidth;
+    const heightRatio = window.innerHeight / canvasHeight;
+    this.setZoom(Math.min(widthRatio, heightRatio));
+    this.setWidth(canvasWidth * this.getZoom());
+    this.setHeight(canvasHeight * this.getZoom());
+  };
+
+  deactivateSelection = async (): Promise<void> => {
+    this.isDrawingMode = false;
+    this.selection = false;
+    this.discardActiveObject();
+    this.forEachObject((object) => {
+      object.selectable = false;
+    });
+    this.requestRenderAll();
+  };
+
+  activateSelection = async (): Promise<void> => {
+    this.isDrawingMode = false;
+    this.selection = true;
+    this.forEachObject((object) => {
+      object.selectable = true;
+    });
+  };
+
+  getNextId = async (): Promise<number> => {
+    this.latestId += 1;
+    return this.latestId;
+  };
+
+  getObjectByIds = (ids: number[]): fabric.Object[] => {
+    // multiple element case; kind of inefficient
+    if (ids.length > 1) {
+      return this.getObjects().filter((object: any) => ids.includes(object.id));
+    }
+    // single element case
+    const id = ids[0];
+    for (let object of this.getObjects()) {
+      if ((object as any).id === id) {
+        return [object];
+      }
+    }
+    return [];
+  };
+
+  serialize = async (objects: fabric.Object[]): Promise<fabric.Object[]> => {
+    return objects.map((object) =>
+      (this as any)._toObject(object, "toObject", ["strokeUniform"])
+    );
+  };
+
+  apply = (ids: number[], newObjects: fabric.Object[] | null): void => {
+    const oldObjects = this.getObjectByIds(ids);
+    if (oldObjects.length) {
+      this.remove(...oldObjects);
+    }
+    if (newObjects && newObjects.length) {
+      fabric.util.enlivenObjects(
+        newObjects,
+        (objects) => {
+          objects.forEach((object: any, i) => {
+            object.id = ids[i];
+          });
+          this.add(...objects);
+        },
+        "fabric"
+      );
+    }
+    this.requestRenderAll();
+  };
+
+  loadFromJSONAsync = async (json: any) =>
+    new Promise<void>((resolve) => {
+      super.loadFromJSON(json, () => {
+        resolve();
+      });
+    });
+}
