+++ src/lib/qboard.ts
@@ -1,7 +1,8 @@
 import { fabric } from "fabric";
 
 import { Handlers, Tool, ToolHandler } from "./tools";
-import { Page, Pages } from "./pages";
+import { Page } from "./page";
+import { Pages } from "./pages";
 import { HistoryHandler } from "./history";
 import { ClipboardHandler } from "./clipboard";
 import { Dash, Fill, Stroke, Style, StyleHandler } from "./styles";
